+++ kartothek/api/dataset.py
@@ -1,4 +1,5 @@
 from kartothek.core import *  # noqa: F401, F403
+from kartothek.core.utils import *  # noqa: F401, F403
 from kartothek.io.dask.bag import *  # noqa: F401, F403
 from kartothek.io.dask.dataframe import *  # noqa: F401, F403
 from kartothek.io.dask.delayed import *  # noqa: F401, F403