+++ kartothek/api/utils.py
@@ -1,1 +1,0 @@
-from kartothek.core.utils import *  # noqa: F401, F403